refactor releaser to adabot and make it pip installable
