--- tools/automated_releaser/README.md.license
@@ -1,3 +1,0 @@
-SPDX-FileCopyrightText: 2023 Tim Cocks
-
-SPDX-License-Identifier: MIT
