--- tools/automated_releaser/clone_bundle.sh
@@ -1,4 +1,0 @@
-git clone https://github.com/adafruit/Adafruit_CircuitPython_Bundle.git
-cd Adafruit_CircuitPython_Bundle
-git submodule sync --quiet --recursive
-git submodule update --init
