--- tools/automated_releaser/release_maker_scripts/templates/release_notes_template.md
@@ -1,5 +1,0 @@
-To use in CircuitPython, simply install the [Adafruit CircuitPython Bundle](https://circuitpython.org/libraries).
-
-To use in CPython, `pip3 install adafruit-circuitpython-{{ pypi_name }}`.
-
-Read the [docs](https://circuitpython.readthedocs.io/projects/{{ pypi_name }}/en/latest/) for info on how to use it.
