--- tools/automated_releaser/release_maker_scripts/templates/release_notes_template.md.license
@@ -1,3 +1,0 @@
-SPDX-FileCopyrightText: 2023 Tim Cocks
-
-SPDX-License-Identifier: MIT
